Update, welcome, README SCHEMA and navbar logout
Just keep working, you need focus and think about how the followers talks with your stars.
They need a link to register and generate a conversation. If another person ingress with the personal link of other user he should see a screen to start a conversation and if the user log througt of the main link he should see yours conversations with others users.